feat: add isBpModalEnabled mixin
